Add initial support for QtTesting
When testing is enabled new menu entries are available, we can pass in
an XML event stream on the command line and replay it. This currently
defaults to existing once replayed, and I added a sample XML event
stream to the data repository. Any XML in the directory are globbed and
added as tests.

We still need to add measurable artifacts, where a simple one for
Avogadro would be image differences or the active molecule after a set
of operations.

Change-Id: I95e35dc1d8cff9846305cfab6cb71780da62f8e5