Added the Open Babel GPLv2 license, ensure installer uses it
When building packages we optionally install the obabel command line
program from Open Babel along with all necessary libraries and data
files. These are licenseed under the GPLv2 (license doesn't currently
seem to be installed). Copy the COPYING file from Open Babel, and if
obabel is being installed combine the GPLv2 with the main BSD license
when installing the package.

Change-Id: I6176d665d1669c50ab6ae3dfee8ce5da2e3127e4