Segregate the edit and general tools
This makes it easier to disable/enable as appropriate. I think we
really need to create separate plugin types and work a little harder
on the UI to ensure only one is active at any given time. This
appraoch at least makes it easy to see what tools are supported for
a given view.

Change-Id: I9e33c370ab18cb3c64112a3b03a122cc6778b622