Added entries for projection type
This works in the basic sense, but still needs some additional work
to set the current projection type when switching views, etc. It is
also begging to be refactored (unless I am getting paid by the line
of code) as the two functions only differ by the enum value. I will
try to get it fixed up soon I promise!

Change-Id: Iebfae5875b27ce924636318f5245c006afab7889