Switched the default to the EditGLWidget
When first opening up the application enable editing. We need to add
some easier/simpler logic to switch widget types (this is not very
intuitive whichever one is default in my opinion).

Ultimately we may want to generalize the widgets to store lists of
editable and "read-only" molecules, with a concept of the active one
which is the one that will be modified. This would let us more simply
switch between different modes, but it is a little more work. Sorry,
what you get here is what I have time to do in the very short term,
I think we can do much better.

Change-Id: I0b1de07961bfb689a504f2925db56dfc5010c576