First pass at singleton with active objects
Uses new class from the Avogadro libraries to manage active objects.
There is more detail in the avogadrolibs repository, but this is a very
limited first use to track global active objects in a running
application. The use of a singleton with active render widgets, etc,
enables the plugins to have a minimal API, and use the singleton API if
they need to gain access to things such as the active render widget to
grab the frame buffer to generate an image, etc.

Signed-off-by: Marcus D. Hanwell <marcus.hanwell@kitware.com>