Parent the instances of plugins, stack mainwindow
This fixes a very old issue I never got around to looking into where the
dtors for plugins were not being called. Problem 1 is that the
MainWindow was allocated on the heap and its destructor wasn't called.
Problem 2 was that we had not parented the plugin instances. This
appears to work as expected now, one plugin was deleting the molecule.

Signed-off-by: Marcus D. Hanwell <marcus@cryos.net>