Rename View Configuration dock to better reflect its function
The name of the View Configuration dock implies that its purpose is to customize the general properties of the view. Perhaps this is once what it did.

However, these days it functions as the place to display extra configuration settings for the display types listed, toggled, and chosen in the Display Types dock.

I believe renaming it to the Display Type Configuration dock better communicates to the user what it does.